Refactor reset-from into separate component
